chore: Remove console.log statements in e2e test file

+++ api/src/__tests__/e2e/books.e2e.ts
@@ -7,7 +7,9 @@ import { createBooksFake } from '@/utils/faker/faker.utils'
 import { EnviromentVariablesConfig } from '@/config/EnviromentVariablesConfig.config'
 
 const dbName = EnviromentVariablesConfig.dbNameTest ?? ''
+console.log('🚀 ~ dbName:', dbName)
 const databaseUrlTest = EnviromentVariablesConfig.databaseUrlTest ?? ''
+console.log('🚀 ~ databaseUrlTest:', databaseUrlTest)
 
 describe('Tests Api', () => {
   let app: Express
@@ -27,12 +29,13 @@ describe('Tests Api', () => {
     })
     await client.connect()
     database = client.db(dbName)
+    console.log('🚀 ~ beforeAll ~ database:', database)
   })
 
   afterAll((doneCallback) => {
     (async () => {
       try {
-        await database.dropDatabase()
+        if (database?.dropDatabase) await database.dropDatabase()
         await client.close()
         server.close(doneCallback)
       } catch (error) {
